requirements.txt
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
cached-property==1.2.0
certifi==2015.9.6.2
coverage==3.7.1
decorator==4.0.4
Django==1.8.4
django-appconf==1.0.1
django-compressor==1.5
django-coverage==1.2.4
django-debug-toolbar==1.4
django-extensions==1.5.7
django-filter==0.11.0
django-friendship==1.3.0
django-modelcluster==0.6.2
django-taggit==0.17.1
django-treebeard==3.0
django-url-filter==0.2.0
djangorestframework==3.2.4
enum34==1.0.4
functools32==3.2.3.post2
html5lib==0.999
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
MarkupSafe==0.23
mistune==0.7.1
nbconvert==4.0.0
nbformat==4.0.1
notebook==4.0.6
path.py==8.1.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.5.2
pydot==1.0.2
Pygments==2.0.2
pyparsing==2.0.3
pytz==2015.6
pyzmq==14.7.0
requests==2.8.1
simplegeneric==0.8.1
six==1.9.0
sqlparse==0.1.16
tornado==4.2.1
traitlets==4.0.0
Unidecode==0.4.18
wagtail==1.1
wheel==0.24.0
Willow==0.2.1
